+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Added Dutch translation.
 - Fixed {% blocktrans %} template whitespace issues
 - Fixed errors when using a column named `items` (#316)
+- Obey `paginate_by` (from `MultipleObjectMixin`) if no later pagination is defined (#242)
 
 ## v1.1.6 (2016-04-02)
 - Correct error message about request context processors for current Django (#314)